refactor(form-builder): use formbuilder onAddField callback
